Refactor code for signup slice
