+++ src/components/SignupForm/Signup.module.css
@@ -62,7 +62,7 @@
   margin: auto;
 }
 
-.formContainer__signIn  {
+.formContainer__signIn {
   text-decoration: underline;
   color: #fff;
   font-weight: bold;
