+++ src/containers/Signup/index.js
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import WaveLoading from 'react-loadingg/lib/WaveLoading';
 
-import { signupUser } from '../../redux/signupSlice';
+import { signupUser } from '../../redux/api/apiActions';
 import SignupForm from '../../components/SignupForm';
 import common from '../../common/common.module.css';
