+++ src/redux/propertySlice.js
@@ -2,7 +2,7 @@
 /* eslint-disable no-param-reassign */
 import axios from 'axios';
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { axiosDefaults, axiosHeders } from '../api/axiosParams';
+import { axiosDefaults, axiosHeders } from './api/axiosParams';
 
 const initialState = {
   loading: false,
